Re-factored the Properties component into two files, set up the properties to be dynamically rendered
